+++ apps/web/src/store/actions/groupViews.js
@@ -328,11 +328,11 @@ export function fetchViewPosts (groupId, viewId) {
 }
 
 /** Create a child space under a parent group. */
-export function createSpace ({ parentGroupId, name, slug, description, icon, acceptedPostTypes, purpose, location, locationId, visibility, accessibility, requiredRoles, viewTypes, bannerUrl, avatarUrl, paywall, addToMenu, status }) {
+export function createSpace ({ parentGroupId, name, slug, description, icon, acceptedPostTypes, purpose, location, locationId, visibility, accessibility, requiredRoles, viewTypes, bannerUrl, avatarUrl, paywall, addToMenu, status, autoAddMembers }) {
   return {
     type: CREATE_SPACE,
     graphql: {
-      query: `mutation ($parentGroupId: ID!, $name: String!, $slug: String, $description: String, $icon: String, $acceptedPostTypes: [String], $purpose: String, $location: String, $locationId: ID, $visibility: Int, $accessibility: Int, $requiredRoles: [Int], $viewTypes: [String], $bannerUrl: String, $avatarUrl: String, $paywall: Boolean, $addToMenu: Boolean, $status: GroupStatus) {
+      query: `mutation ($parentGroupId: ID!, $name: String!, $slug: String, $description: String, $icon: String, $acceptedPostTypes: [String], $purpose: String, $location: String, $locationId: ID, $visibility: Int, $accessibility: Int, $requiredRoles: [Int], $viewTypes: [String], $bannerUrl: String, $avatarUrl: String, $paywall: Boolean, $addToMenu: Boolean, $status: GroupStatus, $autoAddMembers: Boolean) {
         createSpace(
           parentGroupId: $parentGroupId
           name: $name
@@ -352,6 +352,7 @@ export function createSpace ({ parentGroupId, name, slug, description, icon, acc
           paywall: $paywall
           addToMenu: $addToMenu
           status: $status
+          autoAddMembers: $autoAddMembers
         ) {
           id
           name
@@ -366,9 +367,24 @@ export function createSpace ({ parentGroupId, name, slug, description, icon, acc
           accessibility
           status
           active
+          settings {
+            agreementsLastUpdatedAt
+            allowGroupInvites
+            askGroupToGroupJoinQuestions
+            askJoinQuestions
+            defaultDigestFrequency
+            hideExtensionData
+            locationDisplayPrecision
+            publicMemberDirectory
+            showSuggestedSkills
+            showWelcomePage
+            showPostNoticesInChat
+            layout
+            autoAddMembers
+          }
         }
       }`,
-      variables: { parentGroupId, name, slug, description, icon, acceptedPostTypes, purpose, location, locationId, visibility, accessibility, requiredRoles: toIntRoleIds(requiredRoles), viewTypes, bannerUrl, avatarUrl, paywall, addToMenu, status }
+      variables: { parentGroupId, name, slug, description, icon, acceptedPostTypes, purpose, location, locationId, visibility, accessibility, requiredRoles: toIntRoleIds(requiredRoles), viewTypes, bannerUrl, avatarUrl, paywall, addToMenu, status, autoAddMembers }
     },
     meta: {
       parentGroupId,
@@ -380,11 +396,11 @@ export function createSpace ({ parentGroupId, name, slug, description, icon, acc
 }
 
 /** Update a space's settings. */
-export function updateSpace ({ id, groupId, spaceViewId, name, slug, description, icon, acceptedPostTypes, viewName, purpose, location, locationId, visibility, accessibility, requiredRoles, bannerUrl, avatarUrl, paywall, status }) {
+export function updateSpace ({ id, groupId, spaceViewId, name, slug, description, icon, acceptedPostTypes, viewName, purpose, location, locationId, visibility, accessibility, requiredRoles, bannerUrl, avatarUrl, paywall, status, autoAddMembers }) {
   return {
     type: UPDATE_SPACE,
     graphql: {
-      query: `mutation ($id: ID!, $name: String, $slug: String, $description: String, $icon: String, $acceptedPostTypes: [String], $purpose: String, $location: String, $locationId: ID, $visibility: Int, $accessibility: Int, $requiredRoles: [Int], $bannerUrl: String, $avatarUrl: String, $paywall: Boolean, $status: GroupStatus) {
+      query: `mutation ($id: ID!, $name: String, $slug: String, $description: String, $icon: String, $acceptedPostTypes: [String], $purpose: String, $location: String, $locationId: ID, $visibility: Int, $accessibility: Int, $requiredRoles: [Int], $bannerUrl: String, $avatarUrl: String, $paywall: Boolean, $status: GroupStatus, $autoAddMembers: Boolean) {
         updateSpace(
           id: $id
           name: $name
@@ -402,6 +418,7 @@ export function updateSpace ({ id, groupId, spaceViewId, name, slug, description
           avatarUrl: $avatarUrl
           paywall: $paywall
           status: $status
+          autoAddMembers: $autoAddMembers
         ) {
           id
           name
@@ -418,9 +435,24 @@ export function updateSpace ({ id, groupId, spaceViewId, name, slug, description
           icon
           status
           active
+          settings {
+            agreementsLastUpdatedAt
+            allowGroupInvites
+            askGroupToGroupJoinQuestions
+            askJoinQuestions
+            defaultDigestFrequency
+            hideExtensionData
+            locationDisplayPrecision
+            publicMemberDirectory
+            showSuggestedSkills
+            showWelcomePage
+            showPostNoticesInChat
+            layout
+            autoAddMembers
+          }
         }
       }`,
-      variables: omitBy(isUndefined, { id, name, slug, description, icon, acceptedPostTypes, purpose, location, locationId, visibility, accessibility, requiredRoles: toIntRoleIds(requiredRoles), bannerUrl, avatarUrl, paywall, status })
+      variables: omitBy(isUndefined, { id, name, slug, description, icon, acceptedPostTypes, purpose, location, locationId, visibility, accessibility, requiredRoles: toIntRoleIds(requiredRoles), bannerUrl, avatarUrl, paywall, status, autoAddMembers })
     },
     meta: {
       id,