+++ packages/graphql/fragments/membershipFieldsFragment.js
@@ -41,6 +41,7 @@ export default gql`
         showSuggestedSkills
         showWelcomePage
         layout
+        autoAddMembers
       }
       slug
       type